refactor(webui): extract token usage heatmap
